requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
undetected-chromedriver==3.5.4

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
validators==0.22.0

# Web framework (for health check API)
fastapi==0.104.1
uvicorn==0.24.0

# Logging and monitoring
loguru==0.7.2

